requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
mcp>=1.0.0
asyncio-mqtt>=0.16.0
pydantic>=2.0.0
typing-extensions>=4.0.0 
aiohttp